Fix fractional speed integer division truncating to zero in ResolvePriority
IPriorityComparison.Speed was int, causing the fractional speed adjustment
`SpeedOrder.IndexOf(x) / (ActivePerHalf * 2)` to integer-divide to 0 for all
positions. Showdown uses JS float64 numbers, producing correct fractional values
(0.25, 0.5, 0.75) that break ties between same-speed Pokemon during switch-in
event ordering. Changed Speed to double across all IPriorityComparison
implementations and cast the division to floating-point.

+++ ApogeeVGC/Sim/Actions/BeforeTurnAction.cs
@@ -8,7 +8,7 @@ public record BeforeTurnAction : IAction
     public ActionId Choice => ActionId.BeforeTurn;
     public IntFalseUnion Order => 4; // Showdown: orders.beforeTurn = 4
     public double Priority => 0;
-    public int Speed => 0;
+    public double Speed => 0;
     public int SubOrder => 0;
     public int EffectOrder => 0;
     public Pokemon? Pokemon => null;
